+++ airflow/decorators/python.py
@@ -62,6 +62,40 @@ class _PythonDecoratedOperator(DecoratedOperator, PythonOperator):
 T = TypeVar("T", bound=Callable)
 
 
+class PythonDecoratorMixin:
+    """
+    Helper class for inheritance. This class is only used for the __init__.pyi so that IDEs
+    will autocomplete docker decorator functions
+    """
+
+    def __call__(
+        self, python_callable: Optional[Callable] = None, multiple_outputs: Optional[bool] = None, **kwargs
+    ):
+        return self.python(python_callable, multiple_outputs, **kwargs)
+
+    def python(
+        self, python_callable: Optional[Callable] = None, multiple_outputs: Optional[bool] = None, **kwargs
+    ):
+        """
+        Python operator decorator. Wraps a function into an Airflow operator.
+        Accepts kwargs for operator kwarg. Can be reused in a single DAG.
+
+        :param python_callable: Function to decorate
+        :type python_callable: Optional[Callable]
+        :param multiple_outputs: if set, function return value will be
+            unrolled to multiple XCom values. List/Tuples will unroll to xcom values
+            with index as key. Dict will unroll to xcom values with keys as XCom keys.
+            Defaults to False.
+        :type multiple_outputs: bool
+        """
+        return task_decorator_factory(
+            python_callable=python_callable,
+            multiple_outputs=multiple_outputs,
+            decorated_operator_class=_PythonDecoratedOperator,
+            **kwargs,
+        )
+
+
 def python_task(
     python_callable: Optional[Callable] = None, multiple_outputs: Optional[bool] = None, **kwargs
 ):
